AgentSTAExtended: add starting time =0 to disabe the agent MgmtAPExtended: removed relayUnit code. all packets are delivered to upperlayer in raw format. STA does not encapsulate in ethernet anymore. If you use a relayUnit, you need to use an EtherEncap (inverted) to encapsulate the packets.
MgmtSTAExtended:  removed relayUnit code. all packets are delivered to upperlayer in raw format. STA does not encapsulate in ethernet anymore. If you use a relayUnit, you need to use an EtherEncap (inverted) to encapsulate the packets. 
Association info now in inside the AssociatedAPInfo.
Allow to discard packets with no payload.